projects/05-int8-quantization/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tensorrt>=8.5.0
onnx>=1.14.0
nvidia-modelopt>=0.11.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
Pillow>=9.5.0
scipy>=1.10.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=6.5.0
scikit-learn>=1.3.0
pandas>=2.0.0